+++ validation/oneapi_head.jl
@@ -11,7 +11,7 @@ using KernelAbstractions
 using oneAPI
 using Random
 
-const IT = Int64
+const IT = Int32
 const FT = Float32
 const CT = oneAPI.oneArray
 const Backend = oneAPI.oneAPIBackend()
